加水紋效果: 花園的黑色魷魚、最終Boss、石頭加個紅光特效、花園石柱顏色微調

+++ Assets/HelloPico2/Data/VFX/GGDog/Formal/Prefab/HitTarget_TrapPot/Evil_color.shader
@@ -60,26 +60,26 @@ Shader "GGDog/Evil_color"
                 UNITY_VERTEX_INPUT_INSTANCE_ID
             };
             
-            half2 Rotate_UV(half2 uv , half sin , half cos)
+            float2 Rotate_UV(float2 uv , float sin , float cos)
             {
                 return float2(uv.x*cos - uv.y*sin ,uv.x*sin + uv.y*cos);
             }
-            half WaterTex(half2 uv,half Tilling,half FlowSpeed)
+            float WaterTex(float2 uv,float Tilling,float FlowSpeed)
             {
                 uv.xy*=Tilling/50;
-                half Time = _Time.y*FlowSpeed;
+                float Time = _Time.y*FlowSpeed;
 
                 uv.xy = Rotate_UV(uv,0.34,0.14);
-                half2 UV = frac(uv.xy*0.75+Time* half2(-1,-0.25));
-				half D = smoothstep(-10.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                float2 UV = frac(uv.xy*0.75+Time* float2(-1,-0.25));
+				float D = smoothstep(-10.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
                 
                 uv.xy = Rotate_UV(uv,0.94,0.44);
-                UV = frac(uv.xy*1.2+Time*0.33* half2(-0.24,0.33));
-				half D2 = smoothstep(-18.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                UV = frac(uv.xy*1.2+Time*0.33* float2(-0.24,0.33));
+				float D2 = smoothstep(-18.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
                 
                 uv.xy = Rotate_UV(uv,0.64,0.74);
-                UV = frac(uv.xy*1+Time*1.34* half2(0.54,-0.33));
-				half D3 = smoothstep(-15.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                UV = frac(uv.xy*1+Time*1.34* float2(0.54,-0.33));
+				float D3 = smoothstep(-15.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
 
                 D = 1-max(max(D,D2),D3);
                 //D = smoothstep(-3.5,3.5,D+D2+D3);
@@ -181,26 +181,26 @@ Shader "GGDog/Evil_color"
                 UNITY_VERTEX_INPUT_INSTANCE_ID
             };
             
-            half2 Rotate_UV(half2 uv , half sin , half cos)
+            float2 Rotate_UV(float2 uv , float sin , float cos)
             {
                 return float2(uv.x*cos - uv.y*sin ,uv.x*sin + uv.y*cos);
             }
-            half WaterTex(half2 uv,half Tilling,half FlowSpeed)
+            float WaterTex(float2 uv,float Tilling,float FlowSpeed)
             {
                 uv.xy*=Tilling/50;
-                half Time = _Time.y*FlowSpeed;
+                float Time = _Time.y*FlowSpeed;
 
                 uv.xy = Rotate_UV(uv,0.34,0.14);
-                half2 UV = frac(uv.xy*0.75+Time* half2(-1,-0.25));
-				half D = smoothstep(-10.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                float2 UV = frac(uv.xy*0.75+Time* float2(-1,-0.25));
+				float D = smoothstep(-10.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
                 
                 uv.xy = Rotate_UV(uv,0.94,0.44);
-                UV = frac(uv.xy*1.2+Time*0.33* half2(-0.24,0.33));
-				half D2 = smoothstep(-18.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                UV = frac(uv.xy*1.2+Time*0.33* float2(-0.24,0.33));
+				float D2 = smoothstep(-18.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
                 
                 uv.xy = Rotate_UV(uv,0.64,0.74);
-                UV = frac(uv.xy*1+Time*1.34* half2(0.54,-0.33));
-				half D3 = smoothstep(-15.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
+                UV = frac(uv.xy*1+Time*1.34* float2(0.54,-0.33));
+				float D3 = smoothstep(-15.4,4.2,1-38.7*((UV.x-0.5)*(UV.x-0.5)+(UV.y-0.5)*(UV.y-0.5))-1);
 
                 D = 1-max(max(D,D2),D3);
                 //D = smoothstep(-3.5,3.5,D+D2+D3);
